% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_eclipse_data.R
\docType{data}
\name{clean_eclipse_data}
\alias{clean_eclipse_data}
\title{Eclipse Data}
\format{
A data frame with 4,141 rows and 12 variables:
\describe{
\item{Date}{The date of the eclipse (YYYY-MM-DD).}
\item{Type}{The type of eclipse: Total, Annular, Partial, or Hybrid.}
\item{Duration}{The duration of the eclipse in minutes.}
\item{Lat}{The latitude of the eclipse event.}
\item{Lon}{The longitude of the eclipse event.}
}
}
\usage{
data(clean_eclipse_data)
}
\description{
This dataset contains solar eclipses, including their type, date, duration, and location.
}
