% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_eclipses.R
\name{filter_eclipses_by_type}
\alias{filter_eclipses_by_type}
\title{Filter Eclipses by Type}
\usage{
filter_eclipses_by_type(data, eclipse_type)
}
\arguments{
\item{data}{A data frame containing solar eclipse information.}

\item{eclipse_type}{A string specifying the type of eclipse ("Annular 2024" or "Total 2024").}
}
\value{
A data frame filtered by the specified eclipse type.
}
\description{
Filters solar eclipses based on the specified type: "Annular 2024" or "Total 2024".
}
