% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch the Shiny app}
\usage{
launch_app()
}
\value{
Opens a new web window displaying the interactive Shiny app.
}
\description{
This Shiny app allows users to explore solar eclipse data interactively.
Users can filter by eclipse type and minimum duration, visualize the locations on a map,
and view the filtered data in a table.

This function launches the Shiny app included in the package.
}
